Fixes #11
Fixes #11 by adjusted the code by changing the name of the weighted calculator function to calculateRetreivableWelfare and adjusting the logic of the getExtractionTime function to simply return some arbitrary value. The glossary was also updated to fit to the definitions mentioned in the TriagedPatient class.

Note this was previously done in branch #11 but apparently there was an issue with merges